+++ demos/example-electron/vite.renderer.config.ts
@@ -1,10 +1,8 @@
+import react from '@vitejs/plugin-react';
 import type { ConfigEnv, UserConfig } from 'vite';
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import { pluginExposeRenderer } from './vite.base.config.js';
-import wasm from 'vite-plugin-wasm';
-import topLevelAwait from 'vite-plugin-top-level-await';
 import vitePluginRequire from 'vite-plugin-require';
+import { pluginExposeRenderer } from './vite.base.config.js';
 
 // https://vitejs.dev/config
 export default defineConfig((env) => {
@@ -24,22 +22,11 @@ export default defineConfig((env) => {
     optimizeDeps: {
       // Don't optimize these packages as they contain web workers and WASM files.
       // https://github.com/vitejs/vite/issues/11672#issuecomment-1415820673
-      exclude: ['@journeyapps/wa-sqlite', '@powersync/web'],
-      include: []
+      exclude: ['@powersync/web']
     },
-    plugins: [
-      // @ts-expect-error there is TS issue that doesn't actually affect the runtime
-      wasm(),
-      // @ts-expect-error there is TS issue that doesn't actually affect the runtime
-      topLevelAwait(),
-      react(),
-      vitePluginRequire.default(),
-      pluginExposeRenderer(name)
-    ],
+    plugins: [react(), vitePluginRequire.default(), pluginExposeRenderer(name)],
     worker: {
-      format: 'es',
-      // @ts-expect-error there is TS issue that doesn't actually affect the runtime
-      plugins: () => [wasm(), topLevelAwait()]
+      format: 'es'
     },
     resolve: {
       preserveSymlinks: true