+++ demos/example-vite-encryption/vite.config.ts
@@ -1,5 +1,3 @@
-import wasm from 'vite-plugin-wasm';
-import topLevelAwait from 'vite-plugin-top-level-await';
 import { defineConfig } from 'vite';
 
 // https://vitejs.dev/config/
@@ -16,11 +14,9 @@ export default defineConfig({
   optimizeDeps: {
     // Don't optimize these packages as they contain web workers and WASM files.
     // https://github.com/vitejs/vite/issues/11672#issuecomment-1415820673
-    exclude: ['@journeyapps/wa-sqlite', '@powersync/web']
+    exclude: ['@powersync/web']
   },
-  plugins: [wasm(), topLevelAwait()],
   worker: {
-    format: 'es',
-    plugins: () => [wasm(), topLevelAwait()]
+    format: 'es'
   }
 });