+++ packages/adapter-sql-js/vitest.config.ts
@@ -1,19 +1,10 @@
-import topLevelAwait from 'vite-plugin-top-level-await';
-import wasm from 'vite-plugin-wasm';
 import { defineConfig, UserConfigExport } from 'vitest/config';
 
 const config: UserConfigExport = {
   worker: {
-    format: 'es',
-    plugins: () => [wasm(), topLevelAwait()]
+    format: 'es'
   },
-  optimizeDeps: {
-    // Don't optimise these packages as they contain web workers and WASM files.
-    // https://github.com/vitejs/vite/issues/11672#issuecomment-1415820673
-    exclude: ['@journeyapps/wa-sqlite', '@powersync/web'],
-    include: ['async-mutex', 'comlink', 'bson']
-  },
-  plugins: [wasm(), topLevelAwait()],
+  plugins: [],
   test: {
     isolate: false,
     globals: true,