+++ packages/vue/vitest.config.ts
@@ -1,21 +1,17 @@
 import { defineConfig, UserConfigExport } from 'vitest/config';
 
-import topLevelAwait from 'vite-plugin-top-level-await';
-import wasm from 'vite-plugin-wasm';
-
 const config: UserConfigExport = {
   // This is only needed for local tests to resolve the package name correctly
   worker: {
     format: 'es',
-    plugins: () => [wasm(), topLevelAwait()]
+    plugins: () => []
   },
   optimizeDeps: {
     // Don't optimise these packages as they contain web workers and WASM files.
     // https://github.com/vitejs/vite/issues/11672#issuecomment-1415820673
-    exclude: ['@journeyapps/wa-sqlite', '@powersync/web'],
-    include: ['bson', 'async-mutex', 'comlink']
+    exclude: ['@journeyapps/wa-sqlite']
   },
-  plugins: [wasm(), topLevelAwait()],
+  plugins: [],
   test: {
     globals: true,
     include: ['tests/**/*.test.ts'],