+++ tools/diagnostics-app/src/components/navigation/NavigationPage.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { useNavigationPanel } from './NavigationPanelContext';
-import { Box, styled } from '@mui/material';
 
 /**
  * Wraps a component with automatic navigation panel title management
@@ -14,11 +13,5 @@ export const NavigationPage: React.FC<React.PropsWithChildren<{ title: string }>
     return () => navigationPanel.setTitle('');
   }, [title, navigationPanel]);
 
-  return <S.Container>{children}</S.Container>;
+  return <>{children}</>;
 };
-
-namespace S {
-  export const Container = styled(Box)`
-    margin: 10px;
-  `;
-}