+++ packages/capacitor/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/capacitor
 
+## 0.3.0
+
+### Minor Changes
+
+- 25ece59: Improved ESM exports and module declarations. Importing these packages in SSR environments should no longer throw errors.
+
+### Patch Changes
+
+- Updated dependencies [25ece59]
+- Updated dependencies [8db47f3]
+- Updated dependencies [aaf6037]
+- Updated dependencies [acf6b70]
+- Updated dependencies [aaf6037]
+  - @powersync/web@1.32.0
+
 ## 0.2.0
 
 ### Minor Changes